+++ bin/create-user.js
@@ -2,12 +2,19 @@
 
 /**
  * to run create-user applet from the project root folder type in your console:
- * > node bin/create-user <client_id> <client_secret> <email> <password> <first_name> <last_name>
+ * > node bin/create-user <client_id> <client_secret> <email> <password> <first_name> <last_name> <phone_number>
  * <client_id>, <client_secret>, <email>, <password> - are required params
- * <first_name>, <last_name> - are optional
+ * <first_name>, <last_name>, <phone_number> - are optional
+ * options:
+ * --verify-email - send verification email
+ * --start-trial - start 30 day free trial. Defaults to expired subscription
  */
 
 'use strict';
+
+const args = process.argv.slice(2);
+const flags = args.filter(arg => /^--/.test(arg));
+const params = args.filter(arg => !/^--/.test(arg));
 
 const [
   clientId,
@@ -16,7 +23,11 @@ const [
   password,
   first_name,
   last_name,
-] = process.argv.slice(2);
+  number,
+] = params;
+
+const verifyEmail = flags.includes('--verify-email');
+const startTrial = flags.includes('--start-trial');
 
 const { promisify } = require('../utils');
 const { user: { create: createUser } } = require('../lib')({
@@ -31,6 +42,11 @@ createUser$({
   password,
   first_name,
   last_name,
+  number,
+  options: {
+    verifyEmail,
+    startTrial,
+  },
 })
   .then(res => console.log(res))
   .catch(err => console.error(err));